Update assets/inc/class.MCB.php
Hmm. Probably the intent of the original code is to show the meta box when the page cannot be fetched locally so that the user will see the error message.

This is probably what is needed to fix my original issue with the meta box appearing with an erroneous error message for a private custom post type.